Factor out AbstractDelegatedTranspiler
